+++ lib/src/localizations/all.dart
@@ -4,3 +4,5 @@ export './en_US.dart';
 export './mn_MN.dart';
 export './mn_TM.dart';
 export './mn_TM_tn.dart';
+export './ko_KO.dart';
+export './de_DE.dart';